pyproject.toml
[project]
name = "frazer"
description = "Syntatical analysis for sentences in Polish language."
authors = [
    { name = "Adriano Oliveira" }
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dynamic = ["version"]
dependencies = [
    "click==8.*",
    "fastapi==0.115.*",
    "instructor==1.7.2",
    "Jinja2==3.1.*",
    "mangum==0.19.*",
    "openai==1.59.7",
    "pydantic==2.9.*",
    "python-dotenv==1.*",
    "pyyaml>=6.0.2",
    "termcolor>=3.0.1",
    "uvicorn==0.34.*",
]

[project.optional-dependencies]
test = [
    "pytest==8.*",
    "pytest-cov",
]
eval = [
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["frazer*"]

[tool.setuptools.dynamic]
version = {attr = "frazer.__version__"}

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
]
eval = [
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
]

[tool.uv.pip]
no-build = true
no-binary = ["frazer"]
